mv func
